Support ping uplinks
Ping uplinks are supposed to transmit as far as possible. To achieve this we use SF12 and as small a packet size as possible. Pings packets can be transmitted every N transmissions, but using a downlink is preferred. To trigger a transmission of a ping packet, send a downlink containing 0x01. After transmitting the packet, the tracker will reboot to avoid duty cycle restrictions. While this isn't exactly a "nice" solution, it does mean that data packets continue to be transmitted roughly every 30 seconds.